+++ lib/src/id_prefix.rs
@@ -21,6 +21,7 @@ use std::sync::Arc;
 
 use itertools::Itertools as _;
 use once_cell::unsync::OnceCell;
+use thiserror::Error;
 
 use crate::backend::ChangeId;
 use crate::backend::CommitId;
@@ -30,11 +31,19 @@ use crate::object_id::ObjectId;
 use crate::object_id::PrefixResolution;
 use crate::repo::Repo;
 use crate::revset::DefaultSymbolResolver;
+use crate::revset::RevsetEvaluationError;
 use crate::revset::RevsetExpression;
 use crate::revset::RevsetExtensions;
+use crate::revset::RevsetResolutionError;
 use crate::revset::SymbolResolverExtension;
 
-struct PrefixDisambiguationError;
+#[derive(Debug, Error)]
+pub enum IdPrefixIndexLoadError {
+    #[error("Failed to resolve short-prefixes disambiguation revset")]
+    Resolution(#[source] RevsetResolutionError),
+    #[error("Failed to evaluate short-prefixes disambiguation revset")]
+    Evaluation(#[source] RevsetEvaluationError),
+}
 
 struct DisambiguationData {
     expression: Rc<RevsetExpression>,
@@ -52,17 +61,17 @@ impl DisambiguationData {
         &self,
         repo: &dyn Repo,
         extensions: &[impl AsRef<dyn SymbolResolverExtension>],
-    ) -> Result<&Indexes, PrefixDisambiguationError> {
+    ) -> Result<&Indexes, IdPrefixIndexLoadError> {
         self.indexes.get_or_try_init(|| {
             let symbol_resolver = DefaultSymbolResolver::new(repo, extensions);
             let resolved_expression = self
                 .expression
                 .clone()
                 .resolve_user_expression(repo, &symbol_resolver)
-                .map_err(|_| PrefixDisambiguationError)?;
+                .map_err(IdPrefixIndexLoadError::Resolution)?;
             let revset = resolved_expression
                 .evaluate(repo)
-                .map_err(|_| PrefixDisambiguationError)?;
+                .map_err(IdPrefixIndexLoadError::Evaluation)?;
 
             let commit_change_ids = revset.commit_change_ids().collect_vec();
             let mut commit_index = IdIndex::with_capacity(commit_change_ids.len());
@@ -128,14 +137,13 @@ impl IdPrefixContext {
 
     /// Loads disambiguation index once, returns a borrowed index to
     /// disambiguate commit/change IDs.
-    pub fn populate(&self, repo: &dyn Repo) -> IdPrefixIndex<'_> {
-        // TODO: propagate errors instead of treating them as if no revset was specified
-        let indexes = self.disambiguation.as_ref().and_then(|disambiguation| {
-            disambiguation
-                .indexes(repo, self.extensions.symbol_resolvers())
-                .ok()
-        });
-        IdPrefixIndex { indexes }
+    pub fn populate(&self, repo: &dyn Repo) -> Result<IdPrefixIndex<'_>, IdPrefixIndexLoadError> {
+        let indexes = if let Some(disambiguation) = &self.disambiguation {
+            Some(disambiguation.indexes(repo, self.extensions.symbol_resolvers())?)
+        } else {
+            None
+        };
+        Ok(IdPrefixIndex { indexes })
     }
 }